NewChanges_09082022

+++ fcConferenceManager/Controllers/Portolo/PortoloController.cs
@@ -19,7 +19,6 @@ using fcConferenceManager.Models.Portolo;
 using System.Data.SqlClient;
 using OfficeOpenXml;
 using OfficeOpenXml.Table;
-using Microsoft.IdentityModel.Tokens;
 //using PagedList.Mvc;
 //using PagedList;					
 
@@ -31,6 +30,41 @@ namespace fcConferenceManager.Controllers.Portolo
         DBAccessLayer dba = new DBAccessLayer();
         static SqlOperation repository = new SqlOperation();
 
+        bool view;
+        bool add;
+        bool edit;
+        bool delete;
+        int account;
+        public PortoloController()
+        {
+            loginResponse objlt = (loginResponse)System.Web.HttpContext.Current.Session["User"];
+
+            if (objlt != null)
+            {
+                string config = ConfigurationManager.ConnectionStrings["dbconnection"].ConnectionString;
+
+                string query = $"select * from SecurityGroup_Members sm join Privilage_listForPortolo pl on pl.SecurityGroupPkey = sm.SecurityGroup_pKey where sm.Account_pKey = {objlt.Id} and pl.PrivilageID = 'TaskList';";
+
+                using (SqlConnection con = new SqlConnection(config))
+                {
+                    using (SqlCommand cmd = new SqlCommand(query, con))
+                    {
+                        con.Open();
+                        SqlDataReader reader = cmd.ExecuteReader();
+                        while (reader.Read())
+                        {
+                            account = int.Parse(reader["Account_pkey"].ToString());
+                            view = view || bool.Parse(reader["AllowView"].ToString());
+                            add = add || bool.Parse(reader["AllowAdd"].ToString());
+                            edit = edit || bool.Parse(reader["AllowEdit"].ToString());
+                            delete = delete || bool.Parse(reader["AllowDelete"].ToString());
+                        }
+                        reader.Close();
+                        con.Close();
+                    }
+                }
+            }
+        }
         internal static string ReadConnectionString()
         {
 
@@ -89,7 +123,7 @@ namespace fcConferenceManager.Controllers.Portolo
             }
             return selectListItems;
         }
-		private List<SelectListItem> PortoloTaskRepeat_Select_ALL()
+        private List<SelectListItem> PortoloTaskRepeat_Select_ALL()
         {
             List<SelectListItem> selectListItems = new List<SelectListItem>();
             DataTable dt = new DataTable();
@@ -101,12 +135,12 @@ namespace fcConferenceManager.Controllers.Portolo
             }
             return selectListItems;
         }
-														   
+
 
 
         public ActionResult TaskList(List<TaskListResponse> taskListResponse)
         {
-            if (Session["User"] == null)
+            if (Session["User"] == null || !view)
             {
 
                 return Redirect("~/Account/Portolo");
@@ -133,8 +167,9 @@ namespace fcConferenceManager.Controllers.Portolo
                 }
                 else
                 {
-                    TempData.Keep("request");
-                    ViewBag.Request = (TaskListRequest)TempData["request"];
+                    //TempData.Keep("request");
+                    TaskListRequest taskListRequest1 = new TaskListRequest();
+                    ViewBag.Request = taskListRequest1;
 
                 }
                 //List<Response> response = (List<TaskListResponse>)TempData["taskListResponse"];
@@ -158,6 +193,10 @@ namespace fcConferenceManager.Controllers.Portolo
         public ActionResult TaskListfilter(TaskListRequest request)
         {
 
+            if (Session["User"] == null)
+            {
+                Redirect("~/Account/Portolo");
+            }
             TaskListRequest obj = new TaskListRequest();
             request.planType = request.intPortoloRepeat;
 
@@ -238,8 +277,8 @@ namespace fcConferenceManager.Controllers.Portolo
             ViewBag.TaskList = TaskCategoriesList();
             ViewBag.TaskStatus = TaskStatuses_Select_All();
             ViewBag.RepeatTask = TaskRepeat_Select_ALL();
-			            ViewBag.PortoloRepeatTask = PortoloTaskRepeat_Select_ALL();
-					  
+            ViewBag.PortoloRepeatTask = PortoloTaskRepeat_Select_ALL();
+
             ViewBag.Request = obj;
             ViewBag.Baseurl = ConfigurationManager.AppSettings["AppURL"];
             TempData["taskListResponse"] = MainList;
@@ -420,6 +459,7 @@ namespace fcConferenceManager.Controllers.Portolo
 
         public ActionResult AddTopic(Topic model)
         {
+            if (Session["User"] == null) return Redirect("~/Account/Portolo");
             if (model.title == null)
                 return View();
 
@@ -446,6 +486,7 @@ namespace fcConferenceManager.Controllers.Portolo
 
         public ActionResult EditTopic(string id)
         {
+            if (Session["User"] == null) return Redirect("~/Account/Portolo");
             SqlConnection con = new SqlConnection(ReadConnectionString());
 
             string dbquery = " select * from Portolo_Topics where TopicID = " + id;
@@ -467,6 +508,7 @@ namespace fcConferenceManager.Controllers.Portolo
         [HttpPost]
         public ActionResult DeleteTopic(string ids)
         {
+            if (Session["User"] == null) return Redirect("~/Account/Portolo");
             SqlConnection con = new SqlConnection(ReadConnectionString());
 
             string dbquery = String.Format("delete from Portolo_Topics where TopicID in ({0}) ", ids);
@@ -483,9 +525,13 @@ namespace fcConferenceManager.Controllers.Portolo
         [CustomizedAuthorize]
         public ActionResult UserTopics()
         {
+            if (Session["User"] == null)
+            {
+                Redirect("~/Account/Portolo");
+            }
             int Id = ((loginResponse)Session["User"]).Id;
             DataTable dt = new DataTable();
-																																	 
+
             SqlConnection con = new SqlConnection(ReadConnectionString());
 
             string dbquery = String.Format("Select * from Portolo_Topics where IsActive = 1");
@@ -538,9 +584,13 @@ namespace fcConferenceManager.Controllers.Portolo
         }
 
         [HttpGet]
-							 
+
         public ActionResult UserSearchTopic(string name, string description)
         {
+            if (Session["User"] == null)
+            {
+                Redirect("~/Account/Portolo");
+            }
             int Id = ((loginResponse)Session["User"]).Id;
             SqlConnection con = new SqlConnection(ReadConnectionString());
             DataTable dt = new DataTable();
@@ -569,9 +619,13 @@ namespace fcConferenceManager.Controllers.Portolo
 
         public ActionResult SaveUserTopic(string[] ids, string selectall = "")
         {
+            if (Session["User"] == null)
+            {
+                Redirect("~/Account/Portolo");
+            }
             int userId = ((loginResponse)Session["User"]).Id;
             SqlConnection con = new SqlConnection(ReadConnectionString());
-																																		 
+
             con.Open();
             string dbquery = "";
 
@@ -629,6 +683,10 @@ namespace fcConferenceManager.Controllers.Portolo
         public ActionResult commonreload()
         {
 
+            if (Session["User"] == null)
+            {
+                Redirect("~/Account/Portolo");
+            }
             TaskListRequest taskListRequest = new TaskListRequest();
             taskListRequest.active = null;
             taskListRequest.category = "0";
@@ -659,10 +717,36 @@ namespace fcConferenceManager.Controllers.Portolo
 
             List<MyFileUpload> uploadlist = GetFileDetails();
 
+
+
+
+
+
+
+
+
+
+
+
             myMyFileUpload.FileList = uploadlist;
 
             string apppath = GetBaseUrl();
             ViewBag.apppath = apppath;
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
             if (!string.IsNullOrEmpty(search))
             {
@@ -679,6 +763,15 @@ namespace fcConferenceManager.Controllers.Portolo
         {
             MyFileUpload model = new MyFileUpload();
             List<MyFileUpload> list = GetFileDetails();
+
+
+
+
+
+
+
+
+
 
             model.FileList = list;
 
@@ -774,10 +867,11 @@ namespace fcConferenceManager.Controllers.Portolo
             var customer = GetFileDetails().Find(x => x.FileId.Equals(id));
             return Json(customer, JsonRequestBehavior.AllowGet);
         }
-      public ActionResult ProcessLibrary(ProcessLibrary library,string search)
+        public ActionResult ProcessLibrary(ProcessLibrary library, string search)
         {
 
             List<ProcessLibrary> uploadlist = GetProcessDetails();
+
             library.processList = uploadlist;
             if (!string.IsNullOrEmpty(search))
             {
@@ -785,7 +879,8 @@ namespace fcConferenceManager.Controllers.Portolo
                 var searchlist = (from list in uploadlist where list.Process.StartsWith(search.Trim(), StringComparison.OrdinalIgnoreCase) select list).ToList();
                 Session["searchlist"] = searchlist;
                 return View(searchlist);
-            }			 					
+            }
+
             library.processList = uploadlist;
             return View(library.processList);
         }
@@ -796,12 +891,12 @@ namespace fcConferenceManager.Controllers.Portolo
         {
             ProcessLibrary model = new ProcessLibrary();
             List<ProcessLibrary> librarylist = GetProcessDetails();
-            
+
             model.processList = librarylist;
 
             if (Process != null)
             {
-               
+
                 model.Process = Process;
                 if (SaveProcess(model))
                 {
@@ -842,7 +937,7 @@ namespace fcConferenceManager.Controllers.Portolo
         }
         private List<ProcessLibrary> GetProcessDetails()
         {
-            List<ProcessLibrary> librarylist = new List<ProcessLibrary>(); 
+            List<ProcessLibrary> librarylist = new List<ProcessLibrary>();
             string config = ConfigurationManager.ConnectionStrings["dbconnection"].ConnectionString;
             DataTable dtData = new DataTable();
             string query = "Select * from Sys_PortoloProcess";
@@ -863,10 +958,10 @@ namespace fcConferenceManager.Controllers.Portolo
             }
             return librarylist;
         }
-        
+
         public ActionResult DeleteProcess(int pkey)
         {
-            
+
             string config = ConfigurationManager.ConnectionStrings["dbconnection"].ConnectionString;
             using (SqlConnection con = new SqlConnection(config))
             {
@@ -890,12 +985,12 @@ namespace fcConferenceManager.Controllers.Portolo
                         ModelState.Clear();
                     }
 
-                 
+               
                     return Redirect(Request.UrlReferrer.ToString());
                 }
             }
 
-           
+
         }
         public JsonResult EditProcess(int? id)
         {
@@ -904,7 +999,7 @@ namespace fcConferenceManager.Controllers.Portolo
         }
         public ActionResult UpdateProcess(ProcessLibrary library)
         {
-           
+
             string config = ConfigurationManager.ConnectionStrings["dbconnection"].ConnectionString;
             using (SqlConnection con = new SqlConnection(config))
             {
@@ -915,7 +1010,7 @@ namespace fcConferenceManager.Controllers.Portolo
                     cmd.Parameters.AddWithValue("@pkey", library.pkey);
                     cmd.Parameters.AddWithValue("@Process", library.Process.Trim());
                     cmd.Parameters.AddWithValue("@status", "Update");
-                        int result = cmd.ExecuteNonQuery();
+                    int result = cmd.ExecuteNonQuery();
                     if (result == 1)
                     {
                         ViewBag.Message = "Process Updated Successfully";
@@ -927,13 +1022,16 @@ namespace fcConferenceManager.Controllers.Portolo
                         ModelState.Clear();
                     }
                     con.Close();
+
                    
-                
-                   return Redirect(Request.UrlReferrer.ToString());
+                    return Redirect(Request.UrlReferrer.ToString());
+
+
+
 
                 }
             }
-            
+
         }
         public ActionResult DownloadFile(string filePath)
         {
@@ -956,8 +1054,8 @@ namespace fcConferenceManager.Controllers.Portolo
                 throw new System.IO.IOException(s);
             return data;
         }
-	#region ExportExcel
-       public byte[] ExporttoExcel(DataTable table, string filename)
+        #region ExportExcel
+        public byte[] ExporttoExcel(DataTable table, string filename)
         {
 
             ExcelPackage.LicenseContext = LicenseContext.NonCommercial;
@@ -971,9 +1069,7 @@ namespace fcConferenceManager.Controllers.Portolo
         {
             string reportname = $"ProcessLibrary_{DateTime.Now.ToString("yyyyMMddHHmmssfff")}.xlsx";
             var list = GetProcessDetails();
-           
-
-            if (Session["searchlist"]!=null )
+            if (Session["searchlist"]!=null)
             {
                 list = (List<ProcessLibrary>)Session["searchlist"];
 
@@ -983,9 +1079,11 @@ namespace fcConferenceManager.Controllers.Portolo
             dt.Columns.Add("S.No");
             dt.Columns.Add("Process Name");
             
+             
+           
             if (list.Count > 0)
             {
-                foreach(var item in list)
+                foreach (var item in list)
                 {
                     DataRow dataRow = dt.NewRow();
                     dataRow["S.No"] = item.pkey;
@@ -1004,7 +1102,7 @@ namespace fcConferenceManager.Controllers.Portolo
         }
         #endregion					   
 
-        public void DownloadExcel()										   
+        public void DownloadExcel()
         {
             DataTable dt = new DataTable();
             dt.Clear();
@@ -1081,6 +1179,6 @@ namespace fcConferenceManager.Controllers.Portolo
 }
 
 
-    
+
